requirements.txt
anthropic>=0.7.0
openai>=1.0.0
google-generativeai>=0.3.0
rich>=13.0.0
python-dotenv>=1.0.0